backend/requirements.txt
fastapi>=0.114.1,<0.116
httpx>=0.28,<0.29
uvicorn[standard]>=0.36,<0.37
openai>=1.40
openai-chatkit>=1.0.2,<2
